refactor: move log util in common
